requirements.txt
flask==2.2.3
flask-cors==3.0.10
numpy==1.23.5
pillow==9.4.0
torch==1.13.1
torchvision==0.14.1
tritonclient[http]==2.30.0
requests==2.28.2
onnx==1.13.1